:construction: useTalkDataからgameIndexの取得処理を切り出し中
一旦元の動作が保証される形でuseEffectを切り出し
